Fix error throwing in summarizeStreaming()
Previously it was returning rejected promises, but this method does not return promises, so it should throw the exceptions instead.

I considered instead returning an immediately-errored ReadableStream. This seems a bit nicer, but a lot of precedent on the web platform points toward throwing. (Actually, most precedent is for methods returning a promise for a ReadableStream, and returning a rejected promise. I cannot find any methods that currently return a ReadableStream synchronously. But throwing is the sync analogue of returning a rejected promise.)